# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Financemanagment API Documentation",
        "description": "This is auto-generated Swagger API documentation.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a single category",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "operationId": "7dcf71887e585de7f881e9eb98d86126",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Category deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/debts": {
            "get": {
                "tags": [
                    "Debts"
                ],
                "summary": "Get all debts",
                "operationId": "db02cf9efab46d2448020de570b6b6b3",
                "responses": {
                    "200": {
                        "description": "List of debts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DebtResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Debts"
                ],
                "summary": "Create a new debt",
                "operationId": "500be8eb81582c5ff82fa5991dd6a9ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DebtCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Debt created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DebtResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/debts/{id}": {
            "get": {
                "tags": [
                    "Debts"
                ],
                "summary": "Get a single debt",
                "operationId": "54e929e2f8c9527dd45a75fddf38a2c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Debt details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DebtResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Debt not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Debts"
                ],
                "summary": "Update a debt",
                "operationId": "41869a7678b3b2bbd2c866142e9c5493",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DebtUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Debt updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DebtResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Debt not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Debts"
                ],
                "summary": "Delete a debt",
                "operationId": "f84f16bf8eaf35bced9b27b1c18232d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Debt deleted successfully"
                    },
                    "404": {
                        "description": "Debt not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "required": [
                    "id",
                    "user_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Uyga"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CategoryCreate": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Uyga"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "CategoryUpdate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Uyga"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "DebtResource": {
                "title": "Debt",
                "required": [
                    "id",
                    "user_id",
                    "creditor",
                    "amount",
                    "due_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "creditor": {
                        "type": "string",
                        "example": "Ali aka"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-30"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "DebtCreate": {
                "required": [
                    "creditor",
                    "amount",
                    "due_date"
                ],
                "properties": {
                    "creditor": {
                        "type": "string",
                        "example": "Ali aka"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-30"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "DebtUpdate": {
                "properties": {
                    "creditor": {
                        "type": "string",
                        "example": "Ali aka"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-30"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}